adjusted long to long long for MSVC

+++ week_3/day_18/Part_1/day_18.cpp
@@ -7,13 +7,13 @@
 #include"../utils.h"
 
 // forward function declarations
-void snd(const std::string &a, std::vector<int> &freq, std::unordered_map<std::string, long> &reg);
-void set(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg);
-void add(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg);
-void mul(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg);
-void mod(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg);
-void rcv(const std::string &a, std::vector<int> &freq, std::vector<int> &recv, std::unordered_map<std::string, long> &reg);
-void jgz(const std::string &a, const std::string &b, int &i, std::unordered_map<std::string, long> &reg);
+void snd(const std::string &a, std::vector<int> &freq, std::unordered_map<std::string, long long> &reg);
+void set(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg);
+void add(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg);
+void mul(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg);
+void mod(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg);
+void rcv(const std::string &a, std::vector<int> &freq, std::vector<int> &recv, std::unordered_map<std::string, long long> &reg);
+void jgz(const std::string &a, const std::string &b, int &i, std::unordered_map<std::string, long long> &reg);
 
 int main(){
 
@@ -22,7 +22,7 @@ int main(){
     std::vector<std::vector<std::string>> input = read_input_2D("../input_18", delimiters);
 
     // registers 
-    std::unordered_map<std::string, long> reg;
+    std::unordered_map<std::string, long long> reg;
 
     // vector of freq and rcv
     std::vector<int> freq, recv;
@@ -47,36 +47,36 @@ int main(){
     return 0;
 }
 
-void snd(const std::string &a, std::vector<int> &freq, std::unordered_map<std::string, long> &reg){
+void snd(const std::string &a, std::vector<int> &freq, std::unordered_map<std::string, long long> &reg){
     // add new freq
     freq.push_back(reg[a]);
 }
 
-void set(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg){
+void set(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg){
     // check if b is a register or value
     if (b[0]>='a' && b[0]<='z'){ reg[a] = reg[b]; }
     else { reg[a] = std::stoi(b); }
 }
 
-void add(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg){
+void add(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg){
     // check if b is a register or value
     if (b[0]>='a' && b[0]<='z'){ reg[a] += reg[b]; }
     else { reg[a] += std::stoi(b); }
 }
 
-void mul(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg){
+void mul(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg){
     // check if b is a register or value
     if (b[0]>='a' && b[0]<='z'){ reg[a] *= reg[b]; }
     else { reg[a] *= std::stoi(b); }
 }
 
-void mod(const std::string &a, const std::string &b, std::unordered_map<std::string, long> &reg){
+void mod(const std::string &a, const std::string &b, std::unordered_map<std::string, long long> &reg){
     // check if b is a register or value
     if (b[0]>='a' && b[0]<='z'){ reg[a] = mod(reg[a],reg[b]); }
     else { reg[a] = mod(reg[a],std::stol(b)); }
 }
 
-void rcv(const std::string &a, std::vector<int> &freq, std::vector<int> &recv, std::unordered_map<std::string, long> &reg){
+void rcv(const std::string &a, std::vector<int> &freq, std::vector<int> &recv, std::unordered_map<std::string, long long> &reg){
     // check if a is a register or value
     if (a[0]>='a' && a[0]<='z'){ 
         if (reg[a]==0){ return; }
@@ -86,7 +86,7 @@ void rcv(const std::string &a, std::vector<int> &freq, std::vector<int> &recv, s
     recv.push_back(freq.back());
 }
 
-void jgz(const std::string &a, const std::string &b, int &i, std::unordered_map<std::string, long> &reg){
+void jgz(const std::string &a, const std::string &b, int &i, std::unordered_map<std::string, long long> &reg){
     // check if a is a register or value
     if (a[0]>='a' && a[0]<='z'){
         if (reg[a]<=0){ return; }
